initial pass at module for innovation flow state
