+++ src/domain/collaboration/innovation-flow/innovation.flow.module.ts
@@ -9,14 +9,14 @@ import { AuthorizationPolicyModule } from '@domain/common/authorization-policy/a
 import { AuthorizationModule } from '@core/authorization/authorization.module';
 import { ProfileModule } from '@domain/common/profile/profile.module';
 import { TagsetTemplateModule } from '@domain/common/tagset-template/tagset.template.module';
-import { InnovationFlowStatesModule } from '../innovation-flow-states/innovation.flow.state.module';
 import { TagsetModule } from '@domain/common/tagset/tagset.module';
+import { InnovationFlowStateModule } from '../innovation-flow-state/innovation.flow.state.module';
 
 @Module({
   imports: [
     AuthorizationModule,
     AuthorizationPolicyModule,
-    InnovationFlowStatesModule,
+    InnovationFlowStateModule,
     ProfileModule,
     TagsetTemplateModule,
     TagsetModule,
